+++ video2commons/backend/cleanup.sh
