+++ video2commons/config.py
@@ -1,0 +1,1 @@
+../config.py
