--- www/python/src/user-config.py
@@ -1,10 +1,0 @@
-#! /usr/bin/python
-# -*- coding: UTF-8 -*-
-#
-
-"""Pywikibot configs."""
-
-family = 'commons'
-mylang = 'commons'
-
-socket_timeout = 30, 300  # chunked uploading unreliable
